metadata-extraction/requirements.txt
annotated-types==0.7.0
black==24.10.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
flake8==7.1.1
google-api-core==2.24.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-oauth2-tool==0.0.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
kafka-python==2.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
oauth2client==4.1.3
packaging==24.2
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pyflakes==3.2.0
pyparsing==3.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
